feat: add store mod
